Refactor video self view to use <Tile/> component; add CSS for hand raising
